use existing setup file
