+++ tsconfig.json
@@ -9,7 +9,6 @@
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"moduleResolution": "Bundler",
-		"types": ["vitest/globals"]
+		"moduleResolution": "Bundler"
 	}
 }
