+++ vite.config.ts
@@ -11,10 +11,10 @@ export default defineConfig({
 		// in-source testing
 		includeSource: ['src/**/*.{js,ts,svelte}'],
 		// Add @testing-library/jest-dom matchers & mocks of SvelteKit modules
-		setupFiles: ['src/tests/setup.ts'],
+		setupFiles: ['./scripts/setupTest.ts'],
 		// Exclude files in v8
 		coverage: {
-			exclude: ['src/tests/setup.ts']
+			exclude: ['setupTest.ts']
 		},
 		alias: [{ find: /^svelte$/, replacement: 'svelte/internal' }]
 	}
